Work on converting blinky lights to use i2c--add to robot_outputs

+++ robots/claymore/control/lights.h
@@ -50,10 +50,10 @@ struct Lights{
 
 	struct Output{ 
 		bool camera_light;
-		double blinky_light_info;
+		int blinky_light_info;
 
 		Output();
-		Output(bool,double);
+		Output(bool,int);
 	};
 	
 	struct Output_applicator{
